backend/requirements.txt
# 核心依赖
python-dotenv>=1.0.0
pydantic>=2.0.0

# 异步
aiohttp>=3.9.0
aiosqlite>=0.19.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 交易所
ccxt>=4.0.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0

# 工具
python-multipart>=0.0.6
